Refactor chart data handling: remove old data module and implement new data fetching logic with updated structures for historical chart data
